Add handling of venv and freeze to exe to Makefile
This is step 1 in creating a complete solution of handling the
packaging, installing, testing and freezing of python projects.
This includes managing the virtualenv needed for these actions.

This is a generalized solution that should be as project unspecific
as possible. All configuration for a project can then be done in the
settings at the top of the Makefile, so it can be copy-pasted to other
projects easily. The settings can be overridden on a user/system level
by including a `config.mk` if it exists.

This is specifically for use in UNIX-like systems. On Windows, Cygwin
(tested) or WSL (untested) can be used for this purpose. A Windows
specific way of freezing the application to a single executable with
pyinstaller may follow later (or not).