Separate config/db/logs for prod/dev by using ENV
Support setting the base paths with ENV vars. This is needed to have
separate database and config files for production and testing. Before we
even import tagfile, we need to set these to a different value than the
default that is used in production.

When using `make`, these will be automatically set to `/tmp/tagfiletests`
from now on. Using `make test` is now much more prefered to running
`pytest` directly without setting these vars manually.